♻️refactor: 명칭 & 위치 변경
